requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.8
django-anymail==11.0.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.7
inflection==0.5.1
pillow==10.4.0
pycparser==2.22
pyotp==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
